refactor!: move stable plugins from experimental to rolldown/plugins
