+++ packages/rolldown/src/plugins-index.ts
@@ -1,0 +1,2 @@
+export { esmExternalRequirePlugin } from './builtin-plugin/constructors';
+export { replacePlugin } from './builtin-plugin/replace-plugin';
