+++ packages/rolldown/tests/fixtures/builtin-plugin/replace/basic/_config.ts
@@ -1,5 +1,5 @@
 import { defineTest } from 'rolldown-tests';
-import { replacePlugin } from 'rolldown/experimental';
+import { replacePlugin } from 'rolldown/plugins';
 
 export default defineTest({
   config: {
